[refactor] Move tag package outside of build
