+++ pkg/skaffold/tag/git_commit.go
